Add instructions for translating lane id and linear index to matrix coords

+++ crates/cubecl-cpp/src/shared/mma.rs
@@ -53,9 +53,9 @@ pub enum FragmentLayout<D: Dialect> {
 #[derive(Debug, Clone, PartialEq, Eq, Copy)]
 pub struct Fragment<D: Dialect> {
     pub ident: FragmentIdent<D>,
-    pub m: u8,
-    pub n: u8,
-    pub k: u8,
+    pub m: u32,
+    pub n: u32,
+    pub k: u32,
     pub elem: Elem<D>,
     pub layout: Option<FragmentLayout<D>>,
 }
